pyproject.toml
[project]
name = "pharma-state-models-lstm"
version = "0.1.0"
description = "Skeleton for multivariate LSTM forecasting and propensity modeling"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.4",
  "tensorflow>=2.16",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

